--- embedded-foundation/src/bmpdata.rs
@@ -1,4 +1,0 @@
-use tinybmp::Bmp;
-
-pub const RANGE12: &'static [u8] = include_bytes!("../assets/range-12/range-12.bmp");
-pub const RANGE12BASE: &'static [u8] = include_bytes!("../assets/range-12/range-12-base.bmp");
