+++ embedded-foundation/src/modules/state.rs
